fix: reorganize settle expense dialog
